--- README.md
@@ -1,2 +1,0 @@
-# pusling-apps
-Aplikasi admin untuk sistem penjadwalan perpustakaan keliling
