--- jadwal.html
@@ -1,248 +1,0 @@
-<?php 
-
-if (isset($_GET['logout'])) {
-    session_destroy();
-    header('Location: index.php');
-    exit();
-  }
-
-?>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<!-- Boxicons -->
-	<link href='https://unpkg.com/boxicons@2.0.9/css/boxicons.min.css' rel='stylesheet'>
-	<!-- bootstrap -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-	<!-- My CSS -->
-	<link rel="stylesheet" href="css/style.css">
-	<link rel="stylesheet" href="css/responsive.css">
-
-	<title>Pusling</title>
-</head>
-<body>
-
-
-	<!-- SIDEBAR -->
-	<section id="sidebar">
-		<a href="#" class="brand">
-			<i class='bx bxs-truck' ></i>
-			<span class="text">Pusling</span>
-		</a>
-		<ul class="side-menu top">
-			<li>
-				<a href="beranda.php">
-					<i class='bx bxs-dashboard' ></i>
-					<span class="text">Beranda</span>
-				</a>
-			</li>
-			<li class="active">
-				<a href="#">
-					<i class='bx bxs-calendar'></i>
-					<span class="text">Jadwal Kunjungan</span>
-				</a>
-			</li>
-			<li>
-				<a href="kunjungan.php">
-					<i class='bx bxs-truck'></i>
-					<span class="text">Kunjungan</span>
-				</a>
-			</li>
-		</ul>
-		<ul class="side-menu">
-			<li>
-				<a href="pengaturan.php">
-					<i class='bx bxs-cog' ></i>
-					<span class="text">Pengaturan</span>
-				</a>
-			</li>
-			<li>
-				<a href="?logout=true" class="logout">
-					<i class='bx bxs-log-out-circle' ></i>
-					<span class="text">Logout</span>
-				</a>
-			</li>
-		</ul>
-	</section>
-	<!-- SIDEBAR -->
-
-
-
-	<!-- CONTENT -->
-	<section id="content">
-		<!-- NAVBAR -->
-		<nav>
-			<i class='bx bx-menu' ></i>
-			<input type="checkbox" id="switch-mode" hidden>
-			<label for="switch-mode" class="switch-mode"></label>
-		</nav>
-		<!-- NAVBAR -->
-
-		<!-- MAIN -->
-		<main>
-			<div class="head-title">
-				<div class="left">
-					<h1>Jadwal Kunjungan</h1>
-				</div>
-				<a href="#" class="btn-download">
-					<i class='bx bxs-cloud-download' ></i>
-					<span class="text">Download Jadwal</span>
-				</a>
-			</div>
-
-			<!-- <ul class="box-info">
-				<li>
-					<i class='bx bxs-calendar-check' ></i>
-					<span class="text">
-						<h3>10</h3>
-						<p>Total Kunjungan</p>
-					</span>
-				</li>
-				<li>
-					<i class='bx bxs-group' ></i>
-					<span class="text">
-						<h3>2834</h3>
-						<p>Pengunjung</p>
-					</span>
-				</li>
-				<li>
-					<i class='bx bx-history' ></i>
-					<span class="text">
-						<h3>100</h3>
-						<p>Riwayat Kunjungan </p>
-					</span>
-				</li>
-			</ul> -->
-			<section id="calendar">
-				<div class="wrapper">
-					<header>
-						<p class="current-date"></p>
-						<div class="icons">
-							<span id="prev" class="material-symbols-rounded"><i class='bx bx-left-arrow-alt'></i></span>
-							<span id="next" class="material-symbols-rounded"><i class='bx bx-right-arrow-alt'></i></span>
-						</div>
-					</header>
-					<div class="calendar">
-						<ul class="weeks">
-							<li>Minggu</li>
-							<li>Senin</li>
-							<li>Selasa</li>
-							<li>Rabu</li>
-							<li>Kamis</li>
-							<li>Jumat</li>
-							<li>Sabtu</li>
-						</ul>
-						<ul class="days"></ul>
-					</div>
-				</div>
-				
-				<div class="wrapper">
-					<div class="container-todo">
-						<div class="todo">
-							<div class="head">
-								<h3>Jadwal</h3>	
-								<p class="current-date"></p>							
-							</div>
-							<ul class="todo-list">
-								<li class="completed">
-									<p>Tempat Kunjungan</p>
-									<i class='bx bx-dots-vertical-rounded' ></i>
-								</li>
-								<li class="completed">
-									<p>SDN 006 Samarinda</p>
-									<i class='bx bx-dots-vertical-rounded' ></i>
-								</li>
-								<li class="not-completed">
-									<p>SDN 006 Samarinda</p>
-									<i class='bx bx-dots-vertical-rounded' ></i>
-								</li>
-								<li class="completed">
-									<p>SDN 006 Samarinda</p>
-									<i class='bx bx-dots-vertical-rounded' ></i>
-								</li>
-								<li class="not-completed">
-									<p>SDN 006 Samarinda</p>
-									<i class='bx bx-dots-vertical-rounded' ></i>
-								</li>
-							</ul>
-						</div>
-					</div>
-				</div>
-			</section>
-
-			<div class="table-data">
-				<div class="order">
-					<div class="head">
-						<div class="left-head">
-							<h3>Jadwal Kunjungan</h3>
-						</div>
-						<form action="#">
-							<div class="form-input">
-								<input type="search" placeholder="Cari...">
-								<button type="submit" class="search-btn"><i class='bx bx-search' ></i></button>
-							</div>
-						</form>
-						<div class="right-head">
-							<form class="filter">
-								<span>
-									<select>
-										<option value="all">Terbaru</option>
-										<option value="all">Terlama</option>
-									</select>
-								</span>
-								<span>
-									<select>
-										<option value="all">Status</option>
-										<option value="all">Selesai</option>
-										<option value="all">Belum Selesai</option>
-									</select>
-								</span>
-								<span><button>Filter <i class='bx bx-filter' ></i></button></span>
-							</form>
-                        </div>
-					</div>
-					<table>
-						<thead>
-							<tr>
-								<th>Tanggal</th>
-								<th>Jam</th>
-								<th>Sekolah</th>
-								<th>Alamat Sekolah</th>
-								<th>Nomor Hp</th>
-								<th>Status</th>
-							</tr>
-						</thead>
-						<tbody>
-							<tr>
-								<td>01-10-2024</td>
-								<td>05:40</td>
-                                <td>smk negeri 1</td>
-                                <td class="alamat">jl. KH Wahid Hasyim 2 Rt 06 No 21 Samping sungai Keledang</td>
-                                <td>0822222222</td>
-								<td><span class="status completed">Selesai</span></td>
-							</tr>
-							<tr>
-								<td>01-10-2024</td>
-								<td>05:40</td>
-                                <td>smk negeri 1</td>
-                                <td class="alamat">jl. KH Wahid Hasyim 2 Rt 06 No 21 Samping sungai Keledang</td>
-                                <td>0822222222</td>
-								<td><span class="status process">Dalam Proses</span></td>
-							</tr>
-						</tbody>
-					</table>
-				</div>
-			</div>
-		</main>
-		<!-- MAIN -->
-	</section>
-	<!-- CONTENT -->
-	
-
-	<script src="script/script.js"></script>
-	<script src="script/calendar.js"></script>
-</body>
-</html>
