--- login.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Boxicons -->
-    <link href='https://unpkg.com/boxicons@2.0.9/css/boxicons.min.css' rel='stylesheet'>
-    <!-- bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <!-- My CSS -->
-    <link rel="stylesheet" href="css/style.css">
-    <title>Login</title>
-    <style>
-        body {
-            background-color: #eee;
-            font-family: 'Poppins', sans-serif; 
-            height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-        }
-        .login-container {
-            max-width: 400px;
-            margin: 0 auto;
-            padding: 20px;
-            background-color: #fff;
-            border-radius: 8px;
-            box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.075);
-        }
-        .form-group {
-            margin-bottom: 20px;
-        }
-        .form-group input {
-            width: 100%;
-            padding: 10px;
-            border: 1px solid #ccc;
-            border-radius: 5px;
-        }
-        .form-group input:focus {
-            outline: none;
-            border-color: #007bff;
-            box-shadow: 0px 0px 5px 0px rgba(0,123,255,0.3);
-        }
-        .btn-login {
-            width: 100%;
-            padding: 10px;
-            background-color: #007bff;
-            color: #fff;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-        }
-        .btn-login:hover {
-            background-color: #0056b3;
-        }
-        .logo-container {
-            text-align: center;
-            margin-bottom: 20px;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-        }
-        .logo-container i {
-            font-size: 25px;
-            color: #007bff;
-        }
-        .logo-container .text {
-            font-size: 20px;
-            color: #333;
-            margin-left: 10px;
-            vertical-align: middle;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="login-container">
-            <div class="logo-container">
-                <i class='bx bxs-truck'></i><span class="text">Pusling</span>
-            </div>
-            <h2 class="text-center">Login</h2>
-            <form id="login-form" class="needs-validation" novalidate>
-                <div class="form-group">
-                    <input type="text" id="username" name="username" class="form-control" placeholder="Username" required>
-                    <div class="invalid-feedback">
-                        Please enter your username.
-                    </div>
-                </div>
-                <div class="form-group">
-                    <input type="password" id="password" name="password" class="form-control" placeholder="Password" required>
-                    <div class="invalid-feedback">
-                        Please enter your password.
-                    </div>
-                </div>
-                <button type="submit" class="btn btn-primary btn-login">Login</button>
-            </form>
-        </div>
-    </div>
-
-    <script src="script/login.js"></script>        
-</body>
-</html>
