+++ script/calendar.js
@@ -1,60 +1,73 @@
-const daysTag = document.querySelector(".days"),
-    currentDate = document.querySelector(".current-date"),
-    prevNextIcon = document.querySelectorAll(".icons span");
-let date = new Date(),
-    currYear = date.getFullYear(),
-    currMonth = date.getMonth();
+const daysTag = document.querySelector(".days");
+const currentDateElements = document.querySelectorAll(".current-date");
+const prevNextIcon = document.querySelectorAll(".icons span");
+
+let date = new Date();
+let currYear = date.getFullYear();
+let currMonth = date.getMonth();
 
 const daysIndonesian = ["Minggu", "Senin", "Selasa", "Rabu", "Kamis", "Jumat", "Sabtu"];
-// Menambahkan event listener pada setiap elemen tanggal
-daysTag.addEventListener("click", (e) => {
-    if (e.target && e.target.matches("li")) {
-        // Mendapatkan nilai tanggal yang dipilih
-        const selectedDate = parseInt(e.target.textContent);
-        // Memperbarui tampilan kalender sesuai dengan tanggal yang dipilih
-        date.setDate(selectedDate);
-        // Memperbarui kalender
-        renderCalendar();
-    }
-});
-
 const monthsIndonesian = ["Januari", "Februari", "Maret", "April", "Mei", "Juni", "Juli", "Agustus", "September", "Oktober", "November", "Desember"];
 
-
 const renderCalendar = () => {
-    let firstDayofMonth = new Date(currYear, currMonth, 1).getDay(), // getting first day of month
-    lastDateofMonth = new Date(currYear, currMonth + 1, 0).getDate(), // getting last date of month
-    lastDayofMonth = new Date(currYear, currMonth, lastDateofMonth).getDay(), // getting last day of month
-    lastDateofLastMonth = new Date(currYear, currMonth, 0).getDate(); // getting last date of previous month
+    const firstDayOfMonth = new Date(currYear, currMonth, 1).getDay();
+    const lastDateOfMonth = new Date(currYear, currMonth + 1, 0).getDate();
+    const lastDayOfMonth = new Date(currYear, currMonth, lastDateOfMonth).getDay();
+    const lastDateOfLastMonth = new Date(currYear, currMonth, 0).getDate();
+
     let liTag = "";
-    for (let i = firstDayofMonth; i > 0; i--) { // creating li of previous month last days
-        liTag += `<li class="inactive">${lastDateofLastMonth - i + 1}</li>`;
+
+    for (let i = firstDayOfMonth; i > 0; i--) {
+        liTag += `<li class="inactive">${lastDateOfLastMonth - i + 1}</li>`;
     }
-    for (let i = 1; i <= lastDateofMonth; i++) { // creating li of all days of current month
-        // adding active class to li if the current day, month, and year matched
-        let isToday = i === date.getDate() && currMonth === new Date().getMonth() 
-                     && currYear === new Date().getFullYear() ? "active" : "";
+
+    for (let i = 1; i <= lastDateOfMonth; i++) {
+        const isToday = i === date.getDate() && currMonth === new Date().getMonth() && currYear === new Date().getFullYear() ? "active" : "";
         liTag += `<li class="${isToday}">${i}</li>`;
     }
-    for (let i = lastDayofMonth; i < 6; i++) { // creating li of next month first days
-        liTag += `<li class="inactive">${i - lastDayofMonth + 1}</li>`
+
+    for (let i = lastDayOfMonth; i < 6; i++) {
+        liTag += `<li class="inactive">${i - lastDayOfMonth + 1}</li>`;
     }
-    currentDate.innerText = `${daysIndonesian[date.getDay()]}, ${date.getDate()} ${monthsIndonesian[currMonth]} ${currYear}`; // Menampilkan format tanggal aktif dalam bahasa Indonesia
-    daysTag.innerHTML = liTag;
-}
-renderCalendar();
-prevNextIcon.forEach(icon => { // getting prev and next icons
-    icon.addEventListener("click", () => { // adding click event on both icons
-        // if clicked icon is previous icon then decrement current month by 1 else increment it by 1
-        currMonth = icon.id === "prev" ? currMonth - 1 : currMonth + 1;
-        if(currMonth < 0 || currMonth > 11) { // if current month is less than 0 or greater than 11
-            // creating a new date of current year & month and pass it as date value
-            date = new Date(currYear, currMonth, new Date().getDate());
-            currYear = date.getFullYear(); // updating current year with new date year
-            currMonth = date.getMonth(); // updating current month with new date month
-        } else {
-            date = new Date(); // pass the current date as date value
-        }
-        renderCalendar(); // calling renderCalendar function
+
+    currentDateElements.forEach(element => {
+        element.innerText = `${daysIndonesian[date.getDay()]}, ${date.getDate()} ${monthsIndonesian[currMonth]} ${currYear}`;
     });
+
+    daysTag.innerHTML = liTag;
+
+    const formattedDate = `${date.getFullYear()}-${(date.getMonth() + 1).toString().padStart(2, '0')}-${date.getDate().toString().padStart(2, '0')}`;
+    document.getElementById("inputDate").value = formattedDate;
+};
+
+const updateCalendar = () => {
+    if (currMonth < 0 || currMonth > 11) {
+        date = new Date(currYear, currMonth, new Date().getDate());
+        currYear = date.getFullYear();
+        currMonth = date.getMonth();
+    } else {
+        date = new Date();
+    }
+    renderCalendar();
+};
+
+const handleDayClick = (e) => {
+    if (e.target && e.target.matches("li")) {
+        const selectedDate = parseInt(e.target.textContent);
+        date.setDate(selectedDate);
+        renderCalendar();
+    }
+};
+
+const handlePrevNextClick = (icon) => {
+    currMonth = icon.id === "prev" ? currMonth - 1 : currMonth + 1;
+    updateCalendar();
+};
+
+daysTag.addEventListener("click", handleDayClick);
+
+prevNextIcon.forEach(icon => {
+    icon.addEventListener("click", () => handlePrevNextClick(icon));
 });
+
+renderCalendar();
