--- settings.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<!-- Boxicons -->
-	<link href='https://unpkg.com/boxicons@2.0.9/css/boxicons.min.css' rel='stylesheet'>
-	<!-- bootstrap -->
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
-	<!-- My CSS -->
-	<link rel="stylesheet" href="css/style.css">
-	<link rel="stylesheet" href="css/responsive.css">
-
-	<title>Pusling</title>
-</head>
-<body>
-
-
-	<!-- SIDEBAR -->
-	<section id="sidebar">
-		<a href="#" class="brand">
-			<i class='bx bxs-truck' ></i>
-			<span class="text">Pusling</span>
-		</a>
-		<ul class="side-menu top">
-			<li>
-				<a href="index.html">
-					<i class='bx bxs-dashboard' ></i>
-					<span class="text">Beranda</span>
-				</a>
-			</li>
-			<li>
-				<a href="jadwal.html">
-					<i class='bx bxs-calendar'></i>
-					<span class="text">Jadwal Kunjungan</span>
-				</a>
-			</li>
-			<li>
-				<a href="datakunjungan.html">
-					<i class='bx bxs-truck'></i>
-					<span class="text">Data Kunjungan</span>
-				</a>
-			</li>
-		</ul>
-		<ul class="side-menu">
-			<li class="active">
-				<a href="#">
-					<i class='bx bxs-cog' ></i>
-					<span class="text">Pengaturan</span>
-				</a>
-			</li>
-			<li>
-				<a href="#" class="logout">
-					<i class='bx bxs-log-out-circle' ></i>
-					<span class="text">Logout</span>
-				</a>
-			</li>
-		</ul>
-	</section>
-	<!-- SIDEBAR -->
-
-
-
-	<!-- CONTENT -->
-	<section id="content">
-		<!-- NAVBAR -->
-		<nav>
-			<i class='bx bx-menu' ></i>
-			<input type="checkbox" id="switch-mode" hidden>
-			<label for="switch-mode" class="switch-mode"></label>
-		</nav>
-		<!-- NAVBAR -->
-
-		<!-- MAIN -->
-		<main>
-			<div class="head-title">
-				<div class="left">
-					<h1>Pengaturan</h1>
-					<ul class="breadcrumb">
-						<li>
-							<a href="#">Pengaturan</a>
-						</li>
-						<li><i class='bx bx-chevron-right' ></i></li>
-						<li>
-							<a class="active" href="#">Admin</a>
-						</li>
-					</ul>
-				</div>
-			</div>
-
-            <section id="settings">
-                <div class="container">
-                    <div class="row">
-                        <div class="col-md-5">
-                            <form>
-                                <h2>Admin Profile</h2>
-                                <div class="mb-3">
-                                    <label for="username" class="form-label">Username</label>
-                                    <input type="text" class="form-control" id="username" placeholder="Change username">
-                                </div>
-                                <div class="mb-3">
-                                    <label for="password" class="form-label">Password</label>
-                                    <input type="password" class="form-control" id="password" placeholder="Change password">
-                                </div>
-                                <button type="submit" class="btn btn-primary">Save</button>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-            </section>
-
-		</main>
-		<!-- MAIN -->
-	</section>
-	<!-- CONTENT -->
-	
-
-	<script src="script/script.js"></script>
-</body>
-</html>
